# Use caching to improve performance of large profiles