--- a/src/components/views/LocalModList.vue
+++ b/src/components/views/LocalModList.vue
@@ -327,7 +327,7 @@ import SearchUtils from '../../utils/SearchUtils';
         }
 
         getThunderstoreModFromMod(mod: ManifestV2) {
-            return ModBridge.getThunderstoreModFromMod(mod, this.thunderstorePackages);
+            return ModBridge.getCachedThunderstoreModFromMod(mod);
         }
 
         async moveUp(vueMod: any) {
@@ -360,8 +360,8 @@ import SearchUtils from '../../utils/SearchUtils';
             this.filterModList();
         }
 
-        isLatest(vueMod: any): boolean {
-            return ModBridge.isLatestVersion(vueMod);
+        isLatest(mod: ManifestV2): boolean {
+            return ModBridge.isCachedLatestVersion(mod);
         }
 
         getMissingDependencies(vueMod: any): string[] {