@anttimaki 2024-01-10
Forgot to implement the cache clearing when the package list is updated, so changed this to draft.