@anttimaki 2024-01-11
Now clears cache when new package list is downloaded from the API.